Try covariant and upper bound
Upper bound is used incombination with type parameters. When you have some code that accesses the
properties of an object declared using type parameter the compiler have no way to make sure the
accessed property belongs to a particular type. To know this the compiler need a way to understand
to which type the type parameter will belong. We can do this by adding an upper bound. Once the upper
bound is set then any propery belong to the upper bound can be accessed in any of the successor objects
passed as a type parameter.